+++ SurfaceTopography/IO/__init__.py
@@ -47,7 +47,7 @@ from .NPY import NPYReader
 from .PS import PSReader
 from .OPD import OPDReader
 from .OPDx import OPDxReader
-from .VK4 import VK4Reader
+from .VK import VKReader
 from .ZON import ZONReader
 
 # Only writers
@@ -72,7 +72,7 @@ readers = [
     H5Reader,
     NPYReader,
     PSReader,
-    VK4Reader,
+    VKReader,
     ZONReader,
     AL3DReader,
     # HGT reader should come last as there is no file magic