+++ test/IO/test_io.py
@@ -94,8 +94,10 @@ binary_example_file_list = _convert_filelist([
                                                  'example.zon',
                                                  'example.nc',
                                                  'example.vk4',
+                                                 'example.vk6',
                                                  'mitutoyo_mock.xlsx',
-                                                 'mitutoyo_nonuniform_mock.xlsx'
+                                                 'mitutoyo_nonuniform_mock.xlsx',
+                                                 'example_ps.tiff',
                                              ] + [] if NuMPI._has_mpi4py else [
     'example-2d.npy'])  # MPI I/O does not support Python streams
 
@@ -579,10 +581,12 @@ def test_detect_format(file_format_examples):
     assert detect_format(os.path.join(file_format_examples, 'example-2d.npy')) == 'npy'
     assert detect_format(os.path.join(file_format_examples, 'surface.2048x2048.h5')) == 'h5'
     assert detect_format(os.path.join(file_format_examples, 'example.zon')) == 'zon'
-    assert detect_format(os.path.join(file_format_examples, 'example.vk4')) == 'vk4'
+    assert detect_format(os.path.join(file_format_examples, 'example.vk4')) == 'vk'
+    assert detect_format(os.path.join(file_format_examples, 'example.vk6')) == 'vk'
     assert detect_format(os.path.join(file_format_examples, 'mitutoyo_mock.xlsx')) == 'mitutoyo'
     assert detect_format(os.path.join(file_format_examples, 'mitutoyo_nonuniform_mock.xlsx')) == 'mitutoyo'
     assert detect_format(os.path.join(file_format_examples, 'example.al3d')) == 'al3d'
+    assert detect_format(os.path.join(file_format_examples, 'example_ps.tiff')) == 'ps'
 
 
 def test_to_matrix():