+++ SurfaceTopography/ChangeLog.md
@@ -5,6 +5,8 @@ v1.2 (not yet released)
 -----------------------
 
 - ENH: Reader for Park Systems TIFF
+- ENH: Reader for Keyence VK6
+- BUG: Fixed bug in VK4 reader
 
 v1.1 (20Oct22)
 --------------