+++ SurfaceTopography/ChangeLog.md
@@ -5,7 +5,7 @@ v1.2 (not yet released)
 -----------------------
 
 - ENH: Reader for Park Systems TIFF
-- ENH: Reader for Keyence VK6
+- ENH: Reader for Keyence VK3 and VK6
 - MAINT: Removed pandas dependency
 - BUG: Fixed bug in VK4 reader